refactor: split core.py into model.py and workspace.py
Split the monolithic core.py (~1009 lines) into focused modules:
- workspace.py: Workspace class and JSON loading (~393 lines)
- model.py: Model class and tensor compilation (~626 lines, renamed from core.py)

Additional changes:
- Move DomainBounds type alias to typing/aliases.py
- Update __init__.py to import from new canonical locations
- Update all test imports to reference new module paths
  - tests/test_pdf.py: create_bounded_tensor from pyhs3.model
  - tests/test_distributions.py: create_bounded_tensor from pyhs3.model
  - tests/test_normalization_integration.py: Model, Workspace from pyhs3
  - tests/test_unsafe_api.py: Model from pyhs3
- Update documentation to reference new module locations
  - docs/architecture.rst: Update file locations in project structure
  - docs/normalization.rst: Update import examples and class references

No backward-compatibility shim - all imports updated directly.
All tests pass with 99.5% coverage maintained.
