+++ src/pyhs3/__init__.py
@@ -7,8 +7,9 @@ pyhs3: pure-Python HS3 implementation with tensors and autodiff
 from __future__ import annotations
 
 from pyhs3._version import version as __version__
-from pyhs3.core import Model, Workspace
 from pyhs3.logging import setup
+from pyhs3.model import Model
+from pyhs3.workspace import Workspace
 
 setup()
